Add XEN domain cleanup script
Sometimes when a XEN host is managed via libvirt, we can end up in the
situation where a XEN domain is running but it's already gone from
libvirt.

The situation it's logged on libvirt but the message is wrong, as the
message from XEN does not propagate. This script will look for domains
that are in XEN but not in libvirt and shutdown those